geojson-processor/requirements.txt
flask==2.3.3
boto3==1.28.57
psycopg2-binary==2.9.7
geojson==3.0.1
shapely==2.0.1
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0